integer-javaclass
I am still working on this but I still need to learn how to use the millisecond method in Java.  In here I have used the integer method compare to to compare to int type variables that came from the Math.random() class and implemented them in a for loop that prints -1, 1, or 0 depending on the comparison of the two values.  I still need more explanation on this.