integer-javaclass
So here I have worked with a for loop that compares two int type values j and k that are derived from the Math.random() class in Java.  After these values are derived that Integer.compare method compares the two value and return -1, 1 or 0 if j is less than k, or j is greater than k, or if j is equal to k.  I still need to learn the current date and time method from java which I need help with.